requirements.txt
streamlit
pandas
openpyxl
xlsxwriter
cryptography
openai
google-generativeai
langchain
langchain-google-genai
langchain-openai
langchain-core